api: add get_parameters endpoint

+++ bin/docker-run.sh
@@ -1,4 +1,13 @@
 #!/bin/sh
+if [ "z$GOOGLE_APPLICATION_CREDENTIALS" = "z" ]; then
+    printf "ERROR: GOOGLE_APPLICATION_CREDENTIALS not set.\n"
+    exit 1
+fi
+if [ "z$GAR_IMAGE" = "z" ]; then
+   printf "ERROR: GAR_IMAGE not set.\n"
+    exit 1
+fi
+
 docker run --rm -it \
        -p 8000:8000 \
        --env-file .env \
